backend/requirements.txt
fastapi
uvicorn
python-dotenv
playwright
browser_use
langchain_openai
